Поправлен тип гистограммы

+++ crates/intensity/src/histogram.rs
@@ -2,14 +2,14 @@ use image::GrayImage;
 
 const HIST_ARRAY_SIZE: usize = u8::MAX as usize + 1;
 
-pub type HistArray = [f64; HIST_ARRAY_SIZE];
+pub type HistArray = [usize; HIST_ARRAY_SIZE];
 
 #[must_use]
 pub fn histogram(image: &GrayImage) -> HistArray {
-    let mut hist = [0.; HIST_ARRAY_SIZE];
+    let mut hist = [0; HIST_ARRAY_SIZE];
     for sample in image.as_flat_samples().samples {
         if let Some(elem) = hist.get_mut(*sample as usize) {
-            *elem += 1.;
+            *elem += 1;
         }
     }
     hist
@@ -17,5 +17,5 @@ pub fn histogram(image: &GrayImage) -> HistArray {
 
 #[must_use]
 pub fn empty_histogram() -> HistArray {
-    [0.; HIST_ARRAY_SIZE]
+    [0; HIST_ARRAY_SIZE]
 }
